backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
google-genai
python-dotenv
aiofiles
